requirements.txt
streamlit
mcp
groq
httpx
python-dotenv
nest_asyncio
uvicorn